+++ packages/amesos2/src/Amesos2_ShyLUBasker_def.hpp
@@ -295,8 +295,8 @@ ShyLUBasker<Matrix,Vector>::numericFactorization_impl()
       if (ShyLUbasker->Options.dense_schur != 0) {
         if (schur_out_ptr != nullptr) {
           // copy schur out if the output pointer is valid (assuming enough space has been allocated)
-          for (int i = 0; i < schur_size; i++) {
-            for (int j = 0; j < schur_size; j++) {
+          for (size_t i = 0; i < schur_size; i++) {
+            for (size_t j = 0; j < schur_size; j++) {
               schur_out_ptr[i+j*schur_size] = schur_out[i+j*schur_size];
             }
           }
